// src/app/models/reporte.model.ts
export interface Reporte {
  id: number;
  primerNombre: string;
  segundoNombre: string;
  primerApellido: string;
  segundoApellido: string;
  diagnosticoId: number;
  diagnostico: string;
  fechaNacimiento: Date;
  edad: number;
  sexoId: string;
  sexo: null;
  tipoIdentificacionId: string;
  tipoIdentificacion: string;
  numeroIdentificacion: string;
  fechaNotificacionSIVIGILA: Date;
  origenReporteId: number;
  origenReporte: string;
  paisId: string;
  pais: null;
  etniaId: string;
  etnia: string;
  municipioNacimientoId: string;
  departamentoNacimiento: string;
  municipioNacimiento: null;
  grupoPoblacionId: string;
  grupoPoblacion: null;
  residenciaOrigenMunicipioId: string;
  residenciaOrigenDepartamento: null;
  residenciaOrigenMunicipio: null;
  residenciaOrigenBarrio: string;
  residenciaOrigenAreaId: string;
  areaProcedencia: string;
  residenciaOrigenDireccion: string;
  residenciaOrigenEstratoId: string;
  residenciaActualTelefono: string;
  residenciaActualMunicipioId: string;
  municipioResidenciaActual: null;
  departamentoResidenciaActual: null;
  estadoIngresoEstrategiaId: number;
  estadoIngresoEstrategia: string;
  fechaIngresoEstrategia: Date;
  tipoRegimenSSId: string;
  tipoRegimenSS: string;
  epsId: number;
  eps: null;
  ipsId: number;
  ips: null;
  cuidadorNombres: string;
  cuidadorParentescoId: string;
  cuidadorParentesco: string;
  cuidadorEmail: string;
  cuidadorTelefono: string;
  tipoSeguimiento: string;
}